split sync worlflow
