+++ airbyte-workers/src/main/java/io/airbyte/workers/temporal/sync/DbtTransformationActivity.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.workers.temporal.sync;
+
+import io.airbyte.config.OperatorDbtInput;
+import io.airbyte.config.ResourceRequirements;
+import io.airbyte.scheduler.models.IntegrationLauncherConfig;
+import io.airbyte.scheduler.models.JobRunConfig;
+import io.temporal.activity.ActivityInterface;
+import io.temporal.activity.ActivityMethod;
+
+@ActivityInterface
+public interface DbtTransformationActivity {
+
+  @ActivityMethod
+  Void run(JobRunConfig jobRunConfig,
+           IntegrationLauncherConfig destinationLauncherConfig,
+           ResourceRequirements resourceRequirements,
+           OperatorDbtInput input);
+
+}
