+++ airbyte-workers/src/main/java/io/airbyte/workers/temporal/sync/PersistStateActivity.java
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.workers.temporal.sync;
+
+import io.airbyte.config.StandardSyncOutput;
+import io.temporal.activity.ActivityInterface;
+import io.temporal.activity.ActivityMethod;
+import java.util.UUID;
+
+@ActivityInterface
+public interface PersistStateActivity {
+
+  @ActivityMethod
+  boolean persist(final UUID connectionId, final StandardSyncOutput syncOutput);
+
+}
