+++ airbyte-workers/src/main/java/io/airbyte/workers/temporal/sync/SyncWorkflow.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.workers.temporal.sync;
+
+import io.airbyte.config.StandardSyncInput;
+import io.airbyte.config.StandardSyncOutput;
+import io.airbyte.scheduler.models.IntegrationLauncherConfig;
+import io.airbyte.scheduler.models.JobRunConfig;
+import io.temporal.workflow.WorkflowInterface;
+import io.temporal.workflow.WorkflowMethod;
+import java.util.UUID;
+
+@WorkflowInterface
+public interface SyncWorkflow {
+
+  @WorkflowMethod
+  StandardSyncOutput run(JobRunConfig jobRunConfig,
+                         IntegrationLauncherConfig sourceLauncherConfig,
+                         IntegrationLauncherConfig destinationLauncherConfig,
+                         StandardSyncInput syncInput,
+                         UUID connectionId);
+
+}
